+++ examples/multi_agent/include/agentpp_notifytest_mib.h
@@ -65,8 +65,8 @@ public:
 
 	static agentppNotifyTest* instance;
 
-	virtual long       	get_state();
-	virtual void       	set_state(long);
+	virtual int32_t       	get_state();
+	virtual void       	set_state(int32_t);
 	int        	set(const Vbx&) override;
 	int        	prepare_set_request(Request*, int&) override;
 	bool    	value_ok(const Vbx&) override;
